UserContextの作成 及び 編集
UserContextの定義。AppコンポーネントにUserのプロフィール情報取得。子コンポーネントのMainPageで使用する。
